show each coach
